backend/requirements.txt
Flask
python-dotenv
google-generativeai
youtube-transcript-api
fpdf
